Verification API

+++ src/main/java/com/dj/dao/LoginDao.java
@@ -6,9 +6,9 @@ import com.dj.dto.User;
 public interface LoginDao {
 	public User validateLogin(Login login);
 	
-	public boolean invalidateLogin(int userId);
+	public boolean invalidateLogin(Long userId);
 	
-	public boolean userAlreadyLoggedIn(int useId);
+	public boolean userAlreadyLoggedIn(Long useId);
 	
 	public boolean validateToken(String accessToken);
 }
